+++ src/lib/features/feature-toggle/tests/feature-toggle-strategies-store.e2e.test.ts
@@ -1,7 +1,7 @@
 import { IFeatureStrategiesStore } from 'lib/features/feature-toggle/types/feature-toggle-strategies-store-type';
 import { IFeatureToggleStore } from 'lib/features/feature-toggle/types/feature-toggle-store-type';
-import dbInit from '../helpers/database-init';
-import getLogger from '../../fixtures/no-logger';
+import dbInit from '../../../../test/e2e/helpers/database-init';
+import getLogger from '../../../../test/fixtures/no-logger';
 
 let stores;
 let db;
